% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{aveThetas}
\alias{aveThetas}
\title{This computes the AVE and WAVE methods for m-GLMMs}
\usage{
aveThetas(Models, Data, GHk = 5, n, Q, extraParam)
}
\arguments{
\item{Models}{}

\item{Data}{}

\item{GHk}{}

\item{n}{}

\item{Q}{}
}
\value{
The estimates parameters of the model with methods AVE and WAVE
}
\description{
It acts on the Models structure for all pairs of items
}

