% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demoExample.R
\name{demoExample}
\alias{demoExample}
\title{This is a Demo Example of the AVEmGLMM package}
\usage{
demoExample()
}
\value{
Returns the demo example output. It runs as a script with no input.
}
\description{
It sets up an estimation problem, and performs the estimation with methods AVE and WAVE
}

