% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateModelFit.R
\name{estimateModelFit}
\alias{estimateModelFit}
\title{This estimates the modelfit}
\usage{
estimateModelFit(Data, Q, n)
}
\arguments{
\item{Data}{a data.frame with the data. 1st column id, 2nd column time, remaining Q columns are the y 0/1 values (Q items)}

\item{Q}{the number of items. Set this to four.}

\item{n}{number of individuals}
}
\value{
Returns the model fitted
}
\description{
It acts on the Data data.frame and on the number of items Q
}

