% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{This generates the Data}
\usage{
generateData(id, times, n, X, Z, betas, b, Q)
}
\arguments{
\item{id}{}

\item{times}{}

\item{n}{}

\item{X}{}

\item{Z}{}

\item{betas}{}

\item{b}{}

\item{Q}{}
}
\value{
Returns the generated Data as data.frame
}
\description{
It acts on the primitives of the estimation problem and returns the Data
}

